python-service/requirements.txt

fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1

pymilvus>=2.4.0
sentence-transformers==2.5.1
langchain==0.1.9
langchain-community==0.0.24
openai>=1.30.0


python-pptx>=0.6.23
pdfplumber>=0.10.0
python-multipart>=0.0.6
aiofiles>=24.1.0
pytest>=8.0.0
